🐛 Fix a settings behaviour bug
